+++ Mile.Helpers/Mile.Helpers.Base.cpp
@@ -803,3 +803,27 @@ EXTERN_C BOOL WINAPI MileGetFileAttributesByHandle(
 
     return Result;
 }
+
+EXTERN_C BOOL WINAPI MileSetFileAttributesByHandle(
+    _In_ HANDLE FileHandle,
+    _In_ DWORD FileAttributes)
+{
+    FILE_BASIC_INFO BasicInfo = { 0 };
+    BasicInfo.FileAttributes =
+        FileAttributes & (
+            FILE_SHARE_READ |
+            FILE_SHARE_WRITE |
+            FILE_SHARE_DELETE |
+            FILE_ATTRIBUTE_ARCHIVE |
+            FILE_ATTRIBUTE_TEMPORARY |
+            FILE_ATTRIBUTE_OFFLINE |
+            FILE_ATTRIBUTE_NOT_CONTENT_INDEXED |
+            FILE_ATTRIBUTE_NO_SCRUB_DATA) |
+        FILE_ATTRIBUTE_NORMAL;
+
+    return ::SetFileInformationByHandle(
+        FileHandle,
+        FILE_INFO_BY_HANDLE_CLASS::FileBasicInfo,
+        &BasicInfo,
+        sizeof(FILE_BASIC_INFO));
+}